refactor(agent): split oversized AgentChatMessages scroll-follow test and hook into modular pieces (#1658)
